Move transaction listing pages under common component
