+++ packages/frontend/src/pages/expenses/expense.fuctions.tsx
@@ -1,9 +1,9 @@
 import { ExpenseDto } from '@local/types';
 
 import {
-  TransactionStackedListRowProps,
+  TransactionListingItem,
   TransactionType,
-} from '../../components/elements/transaction-stacked-list/transaction-stacked-list.row';
+} from '../../components/blocks/transaction-listing/transaction-listing.item';
 import { formatCurrency } from '../../utils/formatCurrency';
 import { formatDate } from '../../utils/formatDate';
 
@@ -15,13 +15,13 @@ export interface ExpensesPerMonth {
   month: number;
   total: number;
   year: number;
-  rows: TransactionStackedListRowProps[];
+  rows: TransactionListingItem[];
 }
 
 export const convertExpenseToTransactionStackedListRow = (
   expense: ExpenseDto,
   getCategoryName: (id: string) => string | undefined
-): TransactionStackedListRowProps => ({
+): TransactionListingItem => ({
   transactionCategories: expense.categories
     .map(({ category_id }) => getCategoryName(category_id))
     .join(', '),
@@ -43,7 +43,7 @@ export const groupExpensesByMonth = (
 
   const { categoryMappings } = rest;
 
-  const expense: TransactionStackedListRowProps = {
+  const expense: TransactionListingItem = {
     transactionCategories: categoryMappings?.join(', '),
     transactionAmount: formatCurrency(amount),
     date: formatDate(date),
